2021.09.12 Work-flows in progress
This commit's description serves as a sort of continuation of the prior commit's description. Not entirely sure how, but the format of a number of renders/images were corrupted within the repo. The following text represents the names of those files:

- 2021.03.10-VirtualTinyHouse.png
- 2021.04.09-RenderA.png
- 2021.04.09-RenderB.png
- 2021.06.04-RenderA.png
- 2021.06.04-RenderB.png
- 2021.08.07-RenderA.png

Deleted the above files from the repo, located - locally - alternatives to all except the file named 2021.08.07-RenderA.png. Added those alternatives to the repo. Lastly, a large portion of the prior commit's description will remain in effect for perhaps the next three or so commits.